+++ src/backend/core/resource_server/backend.py
@@ -1,6 +1,8 @@
 """Resource Server Backend"""
 
+import json
 import logging
+from json import JSONDecodeError
 
 from django.conf import settings
 from django.contrib import auth
@@ -9,6 +11,7 @@ from django.core.exceptions import ImproperlyConfigured, SuspiciousOperation
 from joserfc import jwe as jose_jwe
 from joserfc import jwt as jose_jwt
 from joserfc.errors import InvalidClaimError, InvalidTokenError
+from joserfc.jwt import Token
 from requests.exceptions import HTTPError
 from rest_framework.exceptions import AuthenticationFailed
 
@@ -50,15 +53,17 @@ class ResourceServerBackend:
             or not authorization_server_client
         ):
             raise ImproperlyConfigured(
-                "Could not instantiate ResourceServerBackend, some parameters are missing."
+                f"Could not instantiate {self.__class__.__name__}: some parameters are missing.",
             )
 
         self._authorization_server_client = authorization_server_client
 
-        self._claims_registry = jose_jwt.JWTClaimsRegistry(
+        self._introspection_claims_registry = jose_jwt.JWTClaimsRegistry(
             iss={"essential": True, "value": self._authorization_server_client.url},
-            aud={"essential": True, "value": self._client_id},
-            token_introspection={"essential": True},
+            active={"essential": True},
+            scope={"essential": True},  # content validated in _verify_user_info
+            # optional in RFC, but required here: "client_id" or "aud"
+            **{settings.OIDC_RS_AUDIENCE_CLAIM: {"essential": True}},
         )
 
         # Declare the token origin audience: to know where the token comes from
@@ -93,7 +98,7 @@ class ResourceServerBackend:
 
         jwt = self._introspect(access_token)
         claims = self._verify_claims(jwt)
-        user_info = self._verify_user_info(claims["token_introspection"])
+        user_info = self._verify_user_info(claims)
 
         sub = user_info.get("sub")
         if sub is None:
@@ -106,7 +111,7 @@ class ResourceServerBackend:
             logger.debug("Login failed: No user with %s found", sub)
             return None
 
-        self.token_origin_audience = str(user_info["aud"])
+        self.token_origin_audience = str(user_info[settings.OIDC_RS_AUDIENCE_CLAIM])
 
         return user
 
@@ -135,7 +140,7 @@ class ResourceServerBackend:
             logger.debug(message)
             raise SuspiciousOperation(message)
 
-        audience = introspection_response.get("aud", None)
+        audience = introspection_response.get(settings.OIDC_RS_AUDIENCE_CLAIM, None)
         if not audience:
             raise SuspiciousOperation(
                 "Introspection response does not provide source audience."
@@ -143,32 +148,42 @@ class ResourceServerBackend:
 
         return introspection_response
 
-    def _introspect(self, token):
-        """Introspect an access token to the authorization server."""
+    def _get_introspection(self, access_token):
+        """Request introspection of an access token to the authorization server."""
         try:
-            jwe = self._authorization_server_client.get_introspection(
-                self._client_id,
-                self._client_secret,
-                token,
+            introspection_response = (
+                self._authorization_server_client.get_introspection(
+                    self._client_id,
+                    self._client_secret,
+                    access_token,
+                )
             )
         except HTTPError as err:
             message = "Could not fetch introspection"
             logger.debug("%s. Exception:", message, exc_info=True)
             raise SuspiciousOperation(message) from err
 
-        private_key = utils.import_private_key_from_settings()
-        jws = self._decrypt(jwe, private_key=private_key)
+        return introspection_response
 
+    def _introspect(self, access_token) -> Token:
+        """
+        Introspect an access token to the authorization server.
+
+        Not implemented here:
+         - introspection_str might be a JWT, not a JSON
+           and therefore should be decoded
+         - introspection_str might be a JWS, not a JSON
+           and therefore should be verified (using self._decode)
+         - introspection_str might be a JWE, not a JSON
+           and therefore should be decrypted (using self._decrypt)
+        """
+        introspection_str = self._get_introspection(access_token)
         try:
-            public_key_set = self._authorization_server_client.import_public_keys()
-        except (TypeError, ValueError, AttributeError, HTTPError) as err:
-            message = "Could get authorization server JWKS"
-            logger.debug("%s. Exception:", message, exc_info=True)
-            raise SuspiciousOperation(message) from err
+            introspection_data = json.loads(introspection_str)
+        except JSONDecodeError as exc:
+            raise SuspiciousOperation("Invalid JSON for introspection") from exc
 
-        jwt = self._decode(jws, public_key_set)
-
-        return jwt
+        return Token({}, introspection_data)
 
     def _decrypt(self, encrypted_token, private_key):
         """Decrypt the token encrypted by the Authorization Server (AS).
@@ -190,7 +205,7 @@ class ResourceServerBackend:
             logger.debug("%s. Exception:", message, exc_info=True)
             raise SuspiciousOperation(message) from err
 
-        return decrypted_token
+        return decrypted_token.plaintext
 
     def _decode(self, encoded_token, public_key_set):
         """Decode the token signed by the Authorization Server (AS).
@@ -201,7 +216,7 @@ class ResourceServerBackend:
         """
         try:
             token = jose_jwt.decode(
-                encoded_token.plaintext,
+                encoded_token,
                 public_key_set,
                 algorithms=[self._signing_algorithm],
             )
@@ -221,13 +236,68 @@ class ResourceServerBackend:
         token substitution or misuse of tokens issued for different clients.
         """
         try:
-            self._claims_registry.validate(token.claims)
+            self._introspection_claims_registry.validate(token.claims)
         except (InvalidClaimError, InvalidTokenError) as err:
             message = "Failed to validate token's claims"
             logger.debug("%s. Exception:", message, exc_info=True)
             raise SuspiciousOperation(message) from err
 
         return token.claims
+
+
+class JWTResourceServerBackend(ResourceServerBackend):
+    """Backend of an OAuth 2.0 resource server.
+
+    Override the classic ResourceServerBackend to support JWT introspection
+    tokens as described in the RFC https://datatracker.ietf.org/doc/rfc9701/
+
+    For this implementation, we expect the introspection response to be
+    in JWT format, signed and encrypted.
+    """
+
+    def _introspect(self, access_token):
+        """
+        Introspect an access token to the authorization server.
+
+        We expect here the `token_introspection` claim to contain the
+        JWT information to be verified:
+         - iss
+         - aud
+         - iat
+
+        Not implemented here:
+         - introspection_str might be a JSON, not a JWE
+         - introspection_str might be a JWS, not a JWE
+        """
+        introspection_str = self._get_introspection(access_token)
+
+        private_key = utils.import_private_key_from_settings()
+        jws = self._decrypt(introspection_str, private_key=private_key)
+
+        try:
+            public_key_set = self._authorization_server_client.import_public_keys()
+        except (TypeError, ValueError, AttributeError, HTTPError) as err:
+            message = "Could get authorization server JWKS"
+            logger.debug("%s. Exception:", message, exc_info=True)
+            raise SuspiciousOperation(message) from err
+
+        jwt = self._decode(jws, public_key_set)
+
+        token_registry = jose_jwt.JWTClaimsRegistry(
+            iss={"essential": True, "value": self._authorization_server_client.url},
+            aud={"essential": True, "value": self._client_id},
+            token_introspection={"essential": True},
+        )
+
+        try:
+            token_registry.validate(jwt.claims)
+        except (InvalidClaimError, InvalidTokenError) as err:
+            logger.exception("JWTResourceServerBackend: %s", err)
+            raise SuspiciousOperation("Failed to validate token's claims") from err
+
+        introspection_data = jwt.claims["token_introspection"]
+
+        return Token({}, introspection_data)
 
 
 class ResourceServerImproperlyConfiguredBackend: